(wip) parsing and validation requests
