+++ client/src/lib.rs
@@ -190,6 +190,7 @@ impl TeeProof {
   pub fn from_manifest(manifest: &Manifest) -> Self {
     let manifest_hash = manifest.to_keccak_digest();
     let data = TeeProofData { manifest_hash: manifest_hash.to_vec() };
+    // TODO: How do I sign this?
     let signature = "sign(hash(TeeProofData))".to_string();
     TeeProof { data, signature }
   }
