+++ notary/src/tee.rs
@@ -12,7 +12,10 @@ use caratls_ekm_server::TeeTlsAcceptor;
 use client::{errors::ClientErrors, origo::OrigoSecrets, TeeProof, TeeProofData};
 use hyper::upgrade::Upgraded;
 use hyper_util::rt::TokioIo;
-use proofs::program::manifest::{Manifest, Response as ManifestResponse};
+use proofs::program::{
+  http::{Request as ManifestRequest, Response as ManifestResponse},
+  manifest::Manifest,
+};
 use serde::Deserialize;
 use tls_client2::tls_core::msgs::message::MessagePayload;
 use tokio::io::{AsyncRead, AsyncReadExt, AsyncWrite, AsyncWriteExt};
@@ -148,8 +151,9 @@ pub async fn tee_proxy_service<S: AsyncWrite + AsyncRead + Send + Unpin>(
   let manifest_bytes = read_wire_struct(&mut tee_tls_stream).await;
   let manifest = Manifest::from_wire_bytes(&manifest_bytes);
   // Checking if manifest is valid
-  // manifest.validate()?; // TODO: This is borking TEE mode, notary just hangs!
+  manifest.validate()?;
   // dbg!(&manifest);
+
   let secret_bytes = read_wire_struct(&mut tee_tls_stream).await;
   let origo_secrets = OrigoSecrets::from_wire_bytes(&secret_bytes);
   // dbg!(&origo_secrets);
@@ -182,27 +186,37 @@ pub async fn tee_proxy_service<S: AsyncWrite + AsyncRead + Send + Unpin>(
 
   let mut app_data_vec = Vec::new();
   for message in &parsed_transcript.payload {
-    if let ParsedMessage { direction, payload, .. } = message {
+    if let ParsedMessage { payload, direction, .. } = message {
       if let Some(app_data) = get_app_data(payload) {
-        if let Ok(readable_data) = String::from_utf8(app_data.clone()) {
-          debug!("{:?} app_data: {}", direction, readable_data);
-        }
+        // if let Ok(readable_data) = String::from_utf8(app_data.clone()) {
+        //   debug!("{:?} app_data: {}", direction, readable_data);
+        // }
         app_data_vec.push(app_data);
       }
     }
   }
-  assert_eq!(app_data_vec.len(), 3);
-  let request = app_data_vec[0].clone();
+
+  if app_data_vec.len() != 3 {
+    return Err(NotaryServerError::MissingAppDataMessages(3, app_data_vec.len()));
+  }
+
+  let request_header = app_data_vec[0].clone();
+  // TODO: Do we expect to get request_body as well part of app_data?
   let response_header = app_data_vec[1].clone();
   let response_body = app_data_vec[2].clone();
+
+  let request = ManifestRequest::from_payload(&request_header, None)?;
+  debug!("{:?}", request);
 
   let response = ManifestResponse::from_payload(&response_header, &response_body)?;
   debug!("{:?}", response);
 
-  if manifest.response.is_subset_of(&response) {
-    debug!("Response matches the manifest");
-  } else {
-    debug!("Response doesn't match the manifest");
+  if !manifest.request.is_subset_of(&request) {
+    return Err(NotaryServerError::ManifestRequestMismatch);
+  }
+
+  if !manifest.response.is_subset_of(&response) {
+    return Err(NotaryServerError::ManifestResponseMismatch);
   }
 
   // send TeeProof to client
@@ -215,15 +229,9 @@ pub async fn tee_proxy_service<S: AsyncWrite + AsyncRead + Send + Unpin>(
 
 fn get_app_data(payload: &WrappedPayload) -> Option<Vec<u8>> {
   match payload {
-    WrappedPayload::Decrypted(decrypted) => {
-      // debug!("decrypted={:?}", decrypted);
-      match &decrypted.payload {
-        MessagePayload::ApplicationData(app_data) => {
-          // debug!("app_data={:?}", app_data);
-          Some(app_data.clone().0)
-        },
-        _ => None,
-      }
+    WrappedPayload::Decrypted(decrypted) => match &decrypted.payload {
+      MessagePayload::ApplicationData(app_data) => Some(app_data.clone().0),
+      _ => None,
     },
     _ => None,
   }
