+++ proofs/src/program/manifest.rs
@@ -49,6 +49,7 @@ use crate::{
   circuits::MAX_STACK_HEIGHT,
   program::{
     data::{CircuitData, FoldInput},
+    http::{Request, Response},
     F, G1,
   },
   ProofError,
@@ -56,9 +57,8 @@ use crate::{
 
 /// HTTP
 const DATA_SIGNAL_NAME: &str = "data";
-
 const PUBLIC_IO_VARS: usize = 11;
-const MAX_HTTP_HEADERS: usize = 25;
+pub(crate) const MAX_HTTP_HEADERS: usize = 25;
 
 /// Manifest containing [`Request`] and [`Response`]
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
@@ -149,349 +149,6 @@ impl Manifest {
     hasher.finalize(&mut output);
 
     output
-  }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub struct ResponseBody {
-  pub json: Vec<JsonKey>,
-}
-
-/// Default HTTP version
-pub fn default_version() -> String { "HTTP/1.1".to_string() }
-/// Default HTTP message
-pub fn default_message() -> String { "OK".to_string() }
-
-/// HTTP Response items required for circuits
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub struct Response {
-  /// HTTP response status
-  pub status:  String,
-  /// HTTP version
-  #[serde(default = "default_version")]
-  pub version: String,
-  /// HTTP response message
-  #[serde(default = "default_message")]
-  pub message: String,
-  /// HTTP headers to lock
-  pub headers: HashMap<String, String>,
-  /// HTTP body keys
-  pub body:    ResponseBody,
-}
-
-impl Response {
-  pub fn validate(&self) -> Result<(), ProofError> {
-    // TODO: What are legal statuses?
-    const VALID_STATUSES: [&str; 2] = ["200", "201"];
-    if !VALID_STATUSES.contains(&self.status.as_str()) {
-      return Err(ProofError::InvalidManifest("Unsupported HTTP status".to_string()));
-    }
-
-    // TODO: What HTTP versions are supported?
-    if self.version != "HTTP/1.1" {
-      return Err(ProofError::InvalidManifest(
-        "Invalid HTTP version: ".to_string() + &self.version,
-      ));
-    }
-
-    // TODO: What is the max supported message length?
-    // TODO: Not covered by serde's #default annotation. Is '""' a valid message?
-    if self.message.len() > 1024 || self.message.is_empty() {
-      return Err(ProofError::InvalidManifest(
-        "Invalid message length: ".to_string() + &self.message,
-      ));
-    }
-
-    // We always expect at least one header, "Content-Type"
-    if self.headers.len() > MAX_HTTP_HEADERS || self.headers.is_empty() {
-      return Err(ProofError::InvalidManifest(
-        "Invalid headers length: ".to_string() + &self.headers.len().to_string(),
-      ));
-    }
-
-    let content_type = self.headers.get("Content-Type");
-    if content_type.is_none() {
-      return Err(ProofError::InvalidManifest("Missing 'Content-Type' header".to_string()));
-    }
-    let content_type = content_type.unwrap();
-
-    const VALID_CONTENT_TYPES: [&str; 2] = ["application/json", "text/plain"];
-    let is_valid_content_type = VALID_CONTENT_TYPES.iter().any(|&legal_type| {
-      content_type == legal_type || content_type.starts_with(&format!("{};", legal_type))
-    });
-    if !is_valid_content_type {
-      return Err(ProofError::InvalidManifest(
-        "Invalid Content-Type header: ".to_string() + content_type,
-      ));
-    }
-
-    // When Content-Type is application/json, we expect at least one JSON item
-    if content_type == "application/json" && self.body.json.is_empty() {
-      return Err(ProofError::InvalidManifest("Expected at least one JSON item".to_string()));
-    }
-
-    Ok(())
-  }
-
-  /// Recovers all `Response` fields from the given payloads and creates a `Response` struct.
-  ///
-  /// - `header_bytes`: The bytes representing the HTTP response headers and metadata.
-  /// - `body_bytes`: The bytes representing the HTTP response body.
-  pub fn from_payload(header_bytes: &[u8], body_bytes: &[u8]) -> Result<Self, ProofError> {
-    let (headers, status, version, message) = Self::parse_header(header_bytes)?;
-    let body = Self::parse_body(body_bytes)?;
-    Ok(Self { status, version, message, headers, body })
-  }
-
-  fn parse_body(body_bytes: &[u8]) -> Result<ResponseBody, ProofError> {
-    let body_json: Value = serde_json::from_slice(body_bytes)
-      .map_err(|_| ProofError::InvalidManifest("Failed to parse body as valid JSON".to_string()))?;
-    let body = ResponseBody {
-      json: match body_json {
-        Value::Object(map) => map.keys().cloned().map(JsonKey::String).collect(),
-        _ =>
-          return Err(ProofError::InvalidManifest("Body is not a valid JSON object".to_string())),
-      },
-    };
-    Ok(body)
-  }
-
-  fn parse_header(
-    header_bytes: &[u8],
-  ) -> Result<(HashMap<String, String>, String, String, String), ProofError> {
-    let headers_str = std::str::from_utf8(header_bytes).map_err(|_| {
-      ProofError::InvalidManifest("Failed to interpret headers as valid UTF-8".to_string())
-    })?;
-    let mut headers = HashMap::new();
-    let mut status = String::new();
-    let mut version = String::new();
-    let mut message = String::new();
-
-    for (i, line) in headers_str.lines().enumerate() {
-      if i == 0 {
-        // Process the first line as the HTTP response start-line (e.g., "HTTP/1.1 200 OK").
-        let parts: Vec<&str> = line.split_whitespace().collect();
-        if parts.len() < 3 {
-          return Err(ProofError::InvalidManifest("Invalid HTTP response start-line".to_string()));
-        }
-        version = parts[0].to_string();
-        status = parts[1].to_string();
-        message = parts[2..].join(" ");
-      } else {
-        // Process subsequent lines as headers (e.g., "Content-Type: application/json").
-        if let Some((key, value)) = line.split_once(": ") {
-          headers.insert(key.to_string(), value.to_string());
-        }
-      }
-    }
-    Ok((headers, status, version, message))
-  }
-
-  /// Performs a test between two response structures. Returns true if `other` contains
-  /// at least all the values also present in `self`.
-  pub fn is_subset_of(&self, other: &Response) -> bool {
-    if self.status != other.status || self.version != other.version || self.message != other.message
-    {
-      debug!("Exact matches don't match");
-      return false;
-    }
-
-    // Check if all headers in `self` are present in `other`
-    for (key, value) in &self.headers {
-      if let Some(other_value) = other.headers.get(key) {
-        if other_value != value {
-          debug!("other_value={} doesnt match value={}", other_value, value);
-          return false;
-        }
-      } else {
-        debug!("missing key {}", key);
-        return false;
-      }
-    }
-
-    // Check if all JSON keys in `self.body` are present in `other.body`
-    for self_key in &self.body.json {
-      if !other.body.json.iter().any(|other_key| self_key == other_key) {
-        debug!("missing self_key {:?}", self_key);
-        return false;
-      }
-    }
-
-    // TODO: We are NOT checking the body contents yet! So this "DSL" is not supported:
-    // 			"body": {
-    // 					"json": [
-    // 						"hello"
-    // 					],
-    // 					"contains": "world"
-    // 				}
-
-    // All checks passed
-    debug!("All checks passed");
-    true
-  }
-}
-
-fn extract_tokens(value: &serde_json::Value) -> Vec<String> {
-  let mut tokens = vec![];
-
-  match value {
-    serde_json::Value::String(s) => {
-      let token_regex = regex::Regex::new(r"<%\s*(\w+)\s*%>").unwrap();
-      for capture in token_regex.captures_iter(s) {
-        if let Some(token) = capture.get(1) {
-          // Extract token name
-          tokens.push(token.as_str().to_string());
-        }
-      }
-    },
-    serde_json::Value::Object(map) => {
-      // Recursively parse nested objects
-      for (_, v) in map {
-        tokens.extend(extract_tokens(v));
-      }
-    },
-    serde_json::Value::Array(arr) => {
-      // Recursively parse arrays
-      for v in arr {
-        tokens.extend(extract_tokens(v));
-      }
-    },
-    _ => {},
-  }
-
-  tokens
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub struct TemplateVar {
-  /// Regex for validation (if applicable)
-  pub regex:  Option<String>,
-  /// Length constraint (if applicable)
-  pub length: Option<usize>,
-  /// Type constraint (e.g., base64, hex)
-  pub r#type: Option<String>,
-}
-
-/// HTTP Request items required for circuits
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub struct Request {
-  /// HTTP method (GET or POST)
-  pub method:  String,
-  /// HTTP request URL
-  pub url:     String,
-  /// HTTP version
-  #[serde(default = "default_version")]
-  pub version: String,
-  /// Request headers to lock
-  pub headers: HashMap<String, String>,
-  /// Request JSON body
-  pub body:    Option<serde_json::Value>,
-  /// Request JSON vars to be used in templates
-  pub vars:    HashMap<String, TemplateVar>,
-}
-
-impl Request {
-  pub fn validate(&self) -> Result<(), ProofError> {
-    // TODO: What HTTP methods are supported?
-    const ALLOWED_METHODS: [&str; 2] = ["GET", "POST"];
-    if !ALLOWED_METHODS.contains(&self.method.as_str()) {
-      return Err(ProofError::InvalidManifest("Invalid HTTP method".to_string()));
-    }
-
-    // Not a valid URL
-    if url::Url::parse(&self.url).is_err() {
-      return Err(ProofError::InvalidManifest("Invalid URL: ".to_string() + &self.url));
-    }
-
-    if !self.url.starts_with("https://") {
-      return Err(ProofError::InvalidManifest("Only HTTPS URLs are allowed".to_string()));
-    }
-
-    // TODO: What HTTP versions are supported?
-    if self.version != "HTTP/1.1" {
-      return Err(ProofError::InvalidManifest(
-        "Invalid HTTP version: ".to_string() + &self.version,
-      ));
-    }
-
-    Ok(())
-  }
-
-  fn validate_vars(&self) -> Result<(), ProofError> {
-    let mut all_tokens = vec![];
-
-    // Parse and validate tokens in the body
-    if let Some(body_tokens) = self.body.as_ref().map(extract_tokens) {
-      for token in &body_tokens {
-        if !self.vars.contains_key(token) {
-          return Err(ProofError::InvalidManifest(format!(
-            "Token `<% {} %>` not declared in `vars`",
-            token
-          )));
-        }
-      }
-      all_tokens.extend(body_tokens);
-    }
-
-    // Parse and validate tokens in headers
-    for value in self.headers.values() {
-      let header_tokens = extract_tokens(&serde_json::Value::String(value.clone()));
-      for token in &header_tokens {
-        if !self.vars.contains_key(token) {
-          return Err(ProofError::InvalidManifest(format!(
-            "Token `<% {} %>` not declared in `vars`",
-            token
-          )));
-        }
-      }
-      all_tokens.extend(header_tokens);
-    }
-
-    for var_key in self.vars.keys() {
-      if !all_tokens.contains(var_key) {
-        return Err(ProofError::InvalidManifest(format!(
-          "Token `<% {} %>` not declared in `body` or `headers`",
-          var_key
-        )));
-      }
-    }
-
-    // Validate each `vars` entry
-    for (key, var_def) in &self.vars {
-      // Validate regex (if defined)
-      if let Some(regex_pattern) = var_def.regex.as_ref() {
-        // Using `regress` crate for compatibility with ECMAScript regular expressions
-        let _regex = regress::Regex::new(regex_pattern).map_err(|_| {
-          ProofError::InvalidManifest(format!("Invalid regex pattern for `{}`", key))
-        })?;
-        // TODO: It will definitely not match it here because it's a template variable, not an
-        // actual variable
-        // TODO: How does the Manifest receiver (notary) verifies this?
-        // if let Some(value) = self.body.as_ref().and_then(|b| b.get(key)) {
-        //   if regex.find(value.as_str().unwrap_or("")).is_none() {
-        //     return Err(ProofError::InvalidManifest(format!(
-        //       "Value for token `<% {} %>` does not match regex",
-        //       key
-        //     )));
-        //   }
-        // }
-      }
-
-      // Validate length (if applicable)
-      if let Some(length) = var_def.length {
-        if let Some(value) = self.body.as_ref().and_then(|b| b.get(key)) {
-          if value.as_str().unwrap_or("").len() != length {
-            return Err(ProofError::InvalidManifest(format!(
-              "Value for token `<% {} %>` does not meet length constraint",
-              key
-            )));
-          }
-        }
-      }
-
-      // TODO: Validate the token "type" constraint
-    }
-    Ok(())
   }
 }
 
@@ -1076,9 +733,13 @@ impl Manifest {
 #[cfg(test)]
 mod tests {
   use super::*;
-  use crate::tests::inputs::{
-    complex_manifest, complex_request_inputs, complex_response_inputs, simple_request_inputs,
-    simple_response_inputs, TEST_MANIFEST,
+  use crate::{
+    create_request, create_response,
+    program::http::{ResponseBody, TemplateVar},
+    tests::inputs::{
+      complex_manifest, complex_request_inputs, complex_response_inputs, simple_request_inputs,
+      simple_response_inputs, TEST_MANIFEST,
+    },
   };
 
   #[test]
@@ -1340,61 +1001,6 @@ mod tests {
     assert_eq!(manifest, wire_deserialized);
   }
 
-  macro_rules! create_request {
-    // Match with optional parameters
-    ($($key:ident: $value:expr),* $(,)?) => {{
-        let mut request = Request {
-            method: "GET".to_string(),
-            url: "https://example.com".to_string(),
-            version: "HTTP/1.1".to_string(),
-            headers: std::collections::HashMap::from([
-                ("Authorization".to_string(), "Bearer <TOKEN>".to_string()),
-                ("User-Agent".to_string(), "test-agent".to_string()),
-            ]),
-            body: None,
-            vars: std::collections::HashMap::from([(
-                "TOKEN".to_string(),
-                TemplateVar {
-                    regex: Some("^[A-Za-z0-9]+$".to_string()),
-                    length: Some(20),
-                    r#type: Some("base64".to_string()),
-                },
-            )]),
-        };
-
-        // Override default fields with provided arguments
-        $(
-            request.$key = $value;
-        )*
-
-        request
-    }};
-  }
-
-  macro_rules! create_response {
-    // Match with optional parameters
-    ($($key:ident: $value:expr),* $(,)?) => {{
-        let mut response = Response {
-            status: "200".to_string(),
-            version: "HTTP/1.1".to_string(),
-            message: "OK".to_string(),
-            headers: std::collections::HashMap::from([
-                ("Content-Type".to_string(), "application/json".to_string())
-            ]),
-            body: ResponseBody {
-                json: vec![JsonKey::String("key1".to_string()), JsonKey::String("key2".to_string())],
-            },
-        };
-
-        // Override default fields with provided arguments
-        $(
-            response.$key = $value;
-        )*
-
-        response
-    }};
-  }
-
   macro_rules! create_manifest {
     (
         $request:expr,
@@ -1545,67 +1151,5 @@ mod tests {
     } else {
       panic!("Expected ProofError::InvalidManifest");
     }
-  }
-
-  #[test]
-  fn test_parse_response() {
-    let header_bytes = b"HTTP/1.1 200 OK\r\nContent-Type: application/json\r\n\r\n";
-    let body_bytes = br#"{"key1": "value1", "key2": "value2"}"#;
-    let response = Response::from_payload(header_bytes, body_bytes).unwrap();
-    assert_eq!(response.status, "200");
-    assert_eq!(response.version, "HTTP/1.1");
-    assert_eq!(response.message, "OK");
-    assert_eq!(response.headers.len(), 1);
-    assert_eq!(response.headers.get("Content-Type").unwrap(), "application/json");
-    assert_eq!(response.body.json.len(), 2);
-    assert_eq!(response.body.json[0], JsonKey::String("key1".to_string()));
-    assert_eq!(response.body.json[1], JsonKey::String("key2".to_string()));
-  }
-
-  #[test]
-  fn test_matches_other_response() {
-    let sourced_response = Response {
-      status:  "200".to_string(),
-      version: "HTTP/1.1".to_string(),
-      message: "OK".to_string(),
-      headers: std::collections::HashMap::from([(
-        "Content-Type".to_string(),
-        "application/json".to_string(),
-      )]),
-      body:    ResponseBody {
-        json: vec![JsonKey::String("key1".to_string()), JsonKey::String("key2".to_string())],
-      },
-    };
-
-    // Matches a perfect match
-    assert!(sourced_response.is_subset_of(&sourced_response));
-
-    // Fails if it doesn't match directly
-    let mut non_matching_response = sourced_response.clone();
-    non_matching_response.status = "201".to_string();
-    assert!(!sourced_response.is_subset_of(&non_matching_response));
-
-    non_matching_response.status = "200".to_string();
-    non_matching_response.message = "Created".to_string();
-    assert!(!sourced_response.is_subset_of(&non_matching_response));
-
-    non_matching_response.message = "OK".to_string();
-    non_matching_response.headers.insert("Content-Type".to_string(), "text/plain".to_string());
-    assert!(!sourced_response.is_subset_of(&non_matching_response));
-
-    // Should pass if the response is a superset of the source response
-    let mut response_superset = sourced_response.clone();
-    response_superset.headers.insert("extra".to_string(), "header".to_string());
-    assert!(sourced_response.is_subset_of(&response_superset));
-
-    let extra_items_in_body = ResponseBody {
-      json: vec![
-        JsonKey::String("key1".to_string()),
-        JsonKey::String("key2".to_string()),
-        JsonKey::String("key3".to_string()),
-      ],
-    };
-    response_superset.body = extra_items_in_body;
-    assert!(sourced_response.is_subset_of(&response_superset));
   }
 }
